requirements.txt
streamlit
torch
torchvision
opencv-python-headless
Pillow
requests
numpy
